requirements.txt
fbchat
emoji
pyyaml
